refactor(commands): extract useInstallMachine hook
source/commands/install.tsx went from 610 lines of mixed state-
machine + presentation down to 184 lines of pure presentation. All
the side-effecting orchestration (resolve, download, parse,
collision detection, backup, render, rollback, hook invocation,
SIGINT handling, runNonInteractive, gate/collision/wizard handlers)
now lives in source/commands/hooks/use-install-machine.ts behind
a single hook.

Exported types (Phase, PreparedContext, HookSummary) live with the
hook so the update command (Milestone 4) can import and extend them
without touching install.tsx.

No behavior change. 142 tests passing, typecheck + build clean.
